[ch72863] event processing cleanup and optimization (#67)
