+++ src/client.c
@@ -4,7 +4,7 @@
 #include <launchdarkly/api.h>
 
 #include "assertion.h"
-#include "events.h"
+#include "event_processor.h"
 #include "network.h"
 #include "client.h"
 #include "config.h"
@@ -38,23 +38,15 @@ LDClientInit(struct LDConfig *const config, const unsigned int maxwaitmilli)
     client->shouldFlush    = false;
     client->shuttingdown   = false;
     client->config         = config;
-    client->summaryStart   = 0;
-    client->lastServerTime = 0;
 
-    LDi_getMonotonicMilliseconds(&client->lastUserKeyFlush);
+    if (!(client->eventProcessor = LDi_newEventProcessor(config))) {
+        LDStoreDestroy(client->store);
+        LDFree(client);
+
+        return NULL;
+    }
+
     LDi_rwlock_init(&client->lock);
-
-    if (!(client->events = LDNewArray())) {
-        goto error;
-    }
-
-    if (!(client->summaryCounters = LDNewObject())) {
-        goto error;
-    }
-
-    if (!(client->userKeys = LDLRUInit(config->userKeysCapacity))) {
-        goto error;
-    }
 
     LDi_thread_create(&client->thread, LDi_networkthread, client);
 
@@ -79,17 +71,6 @@ LDClientInit(struct LDConfig *const config, const unsigned int maxwaitmilli)
     LD_LOG(LD_LOG_INFO, "initialized");
 
     return client;
-
-  error:
-    LDStoreDestroy(client->store);
-
-    LDJSONFree(client->events);
-    LDJSONFree(client->summaryCounters);
-    LDLRUFree(client->userKeys);
-
-    LDFree(client);
-
-    return NULL;
 }
 
 void
@@ -106,9 +87,7 @@ LDClientClose(struct LDClient *const client)
 
         /* cleanup resources */
         LDi_rwlock_destroy(&client->lock);
-        LDJSONFree(client->events);
-        LDJSONFree(client->summaryCounters);
-        LDLRUFree(client->userKeys);
+        LDi_freeEventProcessor(client->eventProcessor);
 
         LDStoreDestroy(client->store);
 
@@ -132,33 +111,11 @@ bool
 LDClientTrack(struct LDClient *const client, const char *const key,
     const struct LDUser *const user, struct LDJSON *const data)
 {
-    struct LDJSON *event, *indexEvent;
-
     LD_ASSERT(client);
     LD_ASSERT(LDUserValidate(user));
     LD_ASSERT(key);
 
-    if (!LDi_maybeMakeIndexEvent(client, user, &indexEvent)) {
-        LD_LOG(LD_LOG_ERROR, "failed to construct index event");
-
-        return false;
-    }
-
-    if (!(event = LDi_newCustomEvent(client, user, key, data))) {
-        LD_LOG(LD_LOG_ERROR, "failed to construct custom event");
-
-        LDJSONFree(indexEvent);
-
-        return false;
-    }
-
-    LDi_addEvent(client, event);
-
-    if (indexEvent) {
-        LDi_addEvent(client, indexEvent);
-    }
-
-    return true;
+    return LDi_track(client->eventProcessor, user, key, data, 0, false);
 }
 
 bool
@@ -166,40 +123,20 @@ LDClientTrackMetric(struct LDClient *const client, const char *const key,
     const struct LDUser *const user, struct LDJSON *const data,
     const double metric)
 {
-    struct LDJSON *event;
-
     LD_ASSERT(client);
     LD_ASSERT(user);
     LD_ASSERT(key);
 
-    if (!(event = LDi_newCustomMetricEvent(client, user, key, data, metric))) {
-        LD_LOG(LD_LOG_ERROR, "failed to construct custom event");
-
-        return false;
-    }
-
-    LDi_addEvent(client, event);
-
-    return true;
+    return LDi_track(client->eventProcessor, user, key, data, metric, true);
 }
 
 bool
 LDClientIdentify(struct LDClient *const client, const struct LDUser *const user)
 {
-    struct LDJSON *event;
-
     LD_ASSERT(client);
     LD_ASSERT(LDUserValidate(user));
 
-    if (!(event = newIdentifyEvent(client, user))) {
-        LD_LOG(LD_LOG_ERROR, "failed to construct identify event");
-
-        return false;
-    }
-
-    LDi_addEvent(client, event);
-
-    return true;
+    return LDi_identify(client->eventProcessor, user);
 }
 
 bool
