+++ src/client.h
@@ -4,6 +4,7 @@
 
 #include "concurrency.h"
 #include "lru.h"
+#include "event_processor.h"
 
 struct LDClient {
     bool initialized;
@@ -11,12 +12,7 @@ struct LDClient {
     struct LDConfig *config;
     ld_thread_t thread;
     ld_rwlock_t lock;
-    struct LDJSON *events; /* Array of Objects */
-    struct LDJSON *summaryCounters; /* Object */
-    unsigned long summaryStart;
     bool shouldFlush;
-    unsigned long long lastServerTime;
-    struct LDLRU *userKeys;
-    unsigned long lastUserKeyFlush;
     struct LDStore *store;
+    struct EventProcessor *eventProcessor;
 };
