+++ src/events.c
@@ -15,167 +15,6 @@
 #include "lru.h"
 
 bool
-LDi_maybeMakeIndexEvent(struct LDClient *const client,
-    const struct LDUser *const user, struct LDJSON **result)
-{
-    unsigned long now;
-    struct LDJSON *event, *tmp;
-    enum LDLRUStatus status;
-
-    LD_ASSERT(client);
-    LD_ASSERT(user);
-    LD_ASSERT(result);
-
-    if (client->config->inlineUsersInEvents) {
-        *result = NULL;
-
-        return true;
-    }
-
-    LDi_getMonotonicMilliseconds(&now);
-
-    LDi_rwlock_wrlock(&client->lock);
-    if (now > client->lastUserKeyFlush +
-        client->config->userKeysFlushInterval) {
-        LDLRUClear(client->userKeys);
-
-        client->lastUserKeyFlush = now;
-    }
-    status = LDLRUInsert(client->userKeys, user->key);
-    LDi_rwlock_wrunlock(&client->lock);
-
-    if (status == LDLRUSTATUS_ERROR) {
-        return false;
-    } else if (status == LDLRUSTATUS_EXISTED) {
-        *result = NULL;
-
-        return true;
-    }
-
-    if (!(event = LDi_newBaseEvent("index"))) {
-        LD_LOG(LD_LOG_ERROR, "alloc error");
-
-        return false;
-    }
-
-    if (!(tmp = LDUserToJSON(client, user, true))) {
-        LD_LOG(LD_LOG_ERROR, "alloc error");
-
-        LDJSONFree(event);
-
-        return false;
-    }
-
-    if (!(LDObjectSetKey(event, "user", tmp))) {
-        LD_LOG(LD_LOG_ERROR, "alloc error");
-
-        LDJSONFree(tmp);
-        LDJSONFree(event);
-
-        return false;
-    }
-
-    *result = event;
-
-    return true;
-}
-
-bool
-LDi_addUserInfoToEvent(struct LDJSON *const event,
-    struct LDClient *const client, const struct LDUser *const user)
-{
-    struct LDJSON *tmp;
-
-    LD_ASSERT(event);
-    LD_ASSERT(client);
-    LD_ASSERT(user);
-
-    if (client->config->inlineUsersInEvents) {
-        if (!(tmp = LDUserToJSON(client, user, true))) {
-            LD_LOG(LD_LOG_ERROR, "alloc error");
-
-            return false;
-        }
-
-        if (!(LDObjectSetKey(event, "user", tmp))) {
-            LD_LOG(LD_LOG_ERROR, "alloc error");
-
-            LDJSONFree(tmp);
-
-            return false;
-        }
-    } else {
-        if (!(tmp = LDNewText(user->key))) {
-            LD_LOG(LD_LOG_ERROR, "alloc error");
-
-            return false;
-        }
-
-        if (!(LDObjectSetKey(event, "userKey", tmp))) {
-            LD_LOG(LD_LOG_ERROR, "alloc error");
-
-            LDJSONFree(tmp);
-
-            return false;
-        }
-    }
-
-    return true;
-}
-
-struct LDJSON *
-LDi_newBaseEvent(const char *const kind)
-{
-    struct LDJSON *tmp, *event;
-    unsigned long milliseconds;
-
-    tmp          = NULL;
-    event        = NULL;
-    milliseconds = 0;
-
-    if (!(event = LDNewObject())) {
-        LD_LOG(LD_LOG_ERROR, "alloc error");
-
-        goto error;
-    }
-
-    LDi_getUnixMilliseconds(&milliseconds);
-
-    if (!(tmp = LDNewNumber(milliseconds))) {
-        LD_LOG(LD_LOG_ERROR, "alloc error");
-
-        goto error;
-    }
-
-    if (!(LDObjectSetKey(event, "creationDate", tmp))) {
-        LD_LOG(LD_LOG_ERROR, "alloc error");
-
-        LDJSONFree(tmp);
-
-        goto error;
-    }
-
-    if (!(tmp = LDNewText(kind))) {
-        LD_LOG(LD_LOG_ERROR, "alloc error");
-
-        goto error;
-    }
-
-    if (!(LDObjectSetKey(event, "kind", tmp))) {
-        LD_LOG(LD_LOG_ERROR, "alloc error");
-
-        goto error;
-    }
-
-    return event;
-
-  error:
-    LDJSONFree(event);
-
-    return NULL;
-}
-
-bool
 LDi_notNull(const struct LDJSON *const json)
 {
     if (json) {
@@ -185,684 +24,6 @@ LDi_notNull(const struct LDJSON *const json)
     }
 
     return false;
-}
-
-struct LDJSON *
-LDi_newFeatureRequestEvent(struct LDClient *const client,
-    const char *const key, const struct LDUser *const user,
-    const unsigned int *const variation, const struct LDJSON *const value,
-    const struct LDJSON *const defaultValue, const char *const prereqOf,
-    const struct LDJSON *const flag, const struct LDDetails *const details)
-{
-    struct LDJSON *tmp, *event;
-
-    LD_ASSERT(key);
-    LD_ASSERT(user);
-
-    tmp   = NULL;
-    event = NULL;
-
-    if (!(event = LDi_newBaseEvent("feature"))) {
-        LD_LOG(LD_LOG_ERROR, "alloc error");
-
-        goto error;
-    }
-
-    if (!LDi_addUserInfoToEvent(event, client, user)) {
-        LD_LOG(LD_LOG_ERROR, "alloc error");
-
-        goto error;
-    }
-
-    if (!(tmp = LDNewText(key))) {
-        LD_LOG(LD_LOG_ERROR, "alloc error");
-
-        goto error;
-    }
-
-    if (!LDObjectSetKey(event, "key", tmp)) {
-        LD_LOG(LD_LOG_ERROR, "alloc error");
-
-        LDJSONFree(tmp);
-
-        goto error;
-    }
-
-    if (variation) {
-        if (!(tmp = LDNewNumber(*variation))) {
-            LD_LOG(LD_LOG_ERROR, "alloc error");
-
-            goto error;
-        }
-
-        if (!LDObjectSetKey(event, "variation", tmp)) {
-            LD_LOG(LD_LOG_ERROR, "alloc error");
-
-            LDJSONFree(tmp);
-
-            goto error;
-        }
-    }
-
-    if (value) {
-        if (!(tmp = LDJSONDuplicate(value))) {
-            LD_LOG(LD_LOG_ERROR, "alloc error");
-
-            goto error;
-        }
-
-        if (!LDObjectSetKey(event, "value", tmp)) {
-            LD_LOG(LD_LOG_ERROR, "alloc error");
-
-            LDJSONFree(tmp);
-
-            goto error;
-        }
-    }
-
-    if (defaultValue) {
-        if (!(tmp = LDJSONDuplicate(defaultValue))) {
-            LD_LOG(LD_LOG_ERROR, "alloc error");
-
-            goto error;
-        }
-
-        if (!LDObjectSetKey(event, "default", tmp)) {
-            LD_LOG(LD_LOG_ERROR, "alloc error");
-
-            LDJSONFree(tmp);
-
-            goto error;
-        }
-    }
-
-    if (prereqOf) {
-        if (!(tmp = LDNewText(prereqOf))) {
-            LD_LOG(LD_LOG_ERROR, "alloc error");
-
-            goto error;
-        }
-
-        if (!LDObjectSetKey(event, "prereqOf", tmp)) {
-            LD_LOG(LD_LOG_ERROR, "alloc error");
-
-            LDJSONFree(tmp);
-
-            goto error;
-        }
-    }
-
-    if (flag) {
-        if (LDi_notNull(tmp = LDObjectLookup(flag, "version"))) {
-            if (LDJSONGetType(tmp) != LDNumber) {
-                LD_LOG(LD_LOG_ERROR, "schema error");
-
-                goto error;
-            }
-
-            if (!(tmp = LDJSONDuplicate(tmp))) {
-                LD_LOG(LD_LOG_ERROR, "memory error");
-
-                goto error;
-            }
-
-            if (!LDObjectSetKey(event, "version", tmp)) {
-                LD_LOG(LD_LOG_ERROR, "memory error");
-
-                LDJSONFree(tmp);
-
-                goto error;
-            }
-        }
-
-        if (LDi_notNull(tmp = LDObjectLookup(flag, "debugEventsUntilDate"))) {
-            if (LDJSONGetType(tmp) != LDNumber) {
-                LD_LOG(LD_LOG_ERROR, "schema error");
-
-                goto error;
-            }
-
-            if (!(tmp = LDJSONDuplicate(tmp))) {
-                LD_LOG(LD_LOG_ERROR, "memory error");
-
-                goto error;
-            }
-
-            if (!LDObjectSetKey(event, "debugEventsUntilDate", tmp)) {
-                LD_LOG(LD_LOG_ERROR, "memory error");
-
-                LDJSONFree(tmp);
-
-                goto error;
-            }
-        }
-
-        if (LDi_notNull(tmp = LDObjectLookup(flag, "trackEvents"))) {
-            if (LDJSONGetType(tmp) != LDBool) {
-                LD_LOG(LD_LOG_ERROR, "schema error");
-
-                goto error;
-            }
-
-            if (LDGetBool(tmp)) {
-                if (!(tmp = LDJSONDuplicate(tmp))) {
-                    LD_LOG(LD_LOG_ERROR, "memory error");
-
-                    goto error;
-                }
-
-                if (!LDObjectSetKey(event, "trackEvents", tmp)) {
-                    LD_LOG(LD_LOG_ERROR, "memory error");
-
-                    LDJSONFree(tmp);
-
-                    goto error;
-                }
-            }
-        }
-    }
-
-    if (details) {
-        if (!(tmp = LDReasonToJSON(details))) {
-            LD_LOG(LD_LOG_ERROR, "memory error");
-
-            goto error;
-        }
-
-        if (!LDObjectSetKey(event, "reason", tmp)) {
-            LD_LOG(LD_LOG_ERROR, "memory error");
-
-            LDJSONFree(tmp);
-
-            goto error;
-        }
-    }
-
-    return event;
-
-  error:
-    LDJSONFree(event);
-
-    return NULL;
-}
-
-struct LDJSON *
-LDi_newCustomEvent(struct LDClient *const client,
-    const struct LDUser *const user, const char *const key,
-    struct LDJSON *const data)
-{
-    struct LDJSON *tmp, *event;
-
-    LD_ASSERT(key);
-
-    tmp   = NULL;
-    event = NULL;
-
-    if (!(event = LDi_newBaseEvent("custom"))) {
-        LD_LOG(LD_LOG_ERROR, "memory error");
-
-        goto error;
-    }
-
-    if (!LDi_addUserInfoToEvent(event, client, user)) {
-        LD_LOG(LD_LOG_ERROR, "alloc error");
-
-        goto error;
-    }
-
-    if (!(tmp = LDNewText(key))) {
-        LD_LOG(LD_LOG_ERROR, "memory error");
-
-        goto error;
-    }
-
-    if (!LDObjectSetKey(event, "key", tmp)) {
-        LD_LOG(LD_LOG_ERROR, "memory error");
-
-        LDJSONFree(tmp);
-
-        goto error;
-    }
-
-    if (data) {
-        if (!LDObjectSetKey(event, "data", data)) {
-            LD_LOG(LD_LOG_ERROR, "memory error");
-
-            goto error;
-        }
-    }
-
-    return event;
-
-  error:
-    LDJSONFree(event);
-
-    return NULL;
-}
-
-struct LDJSON *
-LDi_newCustomMetricEvent(struct LDClient *const client,
-    const struct LDUser *const user, const char *const key,
-    struct LDJSON *const data, const double metric)
-{
-    struct LDJSON *tmp, *event;
-
-    LD_ASSERT(user);
-    LD_ASSERT(key);
-
-    tmp   = NULL;
-    event = NULL;
-
-    if (!(event = LDi_newCustomEvent(client, user, key, data))) {
-        LD_LOG(LD_LOG_ERROR, "memory error");
-
-        goto error;
-    }
-
-    if (!(tmp = LDNewNumber(metric))) {
-        LD_LOG(LD_LOG_ERROR, "memory error");
-
-        goto error;
-    }
-
-    if (!LDObjectSetKey(event, "metricValue", tmp)) {
-        LD_LOG(LD_LOG_ERROR, "memory error");
-
-        LDJSONFree(tmp);
-
-        goto error;
-    }
-
-    return event;
-
-  error:
-    LDJSONFree(event);
-
-    return NULL;
-}
-
-struct LDJSON *
-newIdentifyEvent(struct LDClient *const client, const struct LDUser *const user)
-{
-    struct LDJSON *event, *tmp;
-
-    LD_ASSERT(client);
-    LD_ASSERT(LDUserValidate(user));
-
-    event = NULL;
-    tmp   = NULL;
-
-    if (!(event = LDi_newBaseEvent("identify"))) {
-        LD_LOG(LD_LOG_ERROR, "failed to construct base event");
-
-        return NULL;
-    }
-
-    if (!(tmp = LDNewText(user->key))) {
-        LD_LOG(LD_LOG_ERROR, "alloc error");
-
-        LDJSONFree(event);
-
-        return false;
-    }
-
-    if (!(LDObjectSetKey(event, "key", tmp))) {
-        LD_LOG(LD_LOG_ERROR, "alloc error");
-
-        LDJSONFree(event);
-        LDJSONFree(tmp);
-
-        return false;
-    }
-
-    if (!(tmp = LDUserToJSON(client, user, true))) {
-        LD_LOG(LD_LOG_ERROR, "alloc error");
-
-        LDJSONFree(event);
-
-        return NULL;
-    }
-
-    if (!(LDObjectSetKey(event, "user", tmp))) {
-        LD_LOG(LD_LOG_ERROR, "alloc error");
-
-        LDJSONFree(tmp);
-        LDJSONFree(event);
-
-        return NULL;
-    }
-
-    return event;
-}
-
-void
-LDi_addEvent(struct LDClient *const client, struct LDJSON *const event)
-{
-    LD_ASSERT(client);
-    LD_ASSERT(event);
-
-    LDi_rwlock_wrlock(&client->lock);
-
-    /* sanity check */
-    LD_ASSERT(LDJSONGetType(client->events) == LDArray);
-
-    if (LDCollectionGetSize(client->events) >= client->config->eventsCapacity) {
-        LD_LOG(LD_LOG_WARNING, "event capacity exceeded, dropping event");
-
-        LDi_rwlock_wrunlock(&client->lock);
-    } else {
-        LDArrayPush(client->events, event);
-
-        LDi_rwlock_wrunlock(&client->lock);
-    }
-}
-
-char *
-LDi_makeSummaryKey(const struct LDJSON *const event)
-{
-    char *keytext;
-    struct LDJSON *key, *tmp;
-
-    LD_ASSERT(event);
-
-    tmp     = NULL;
-    keytext = NULL;
-    key     = NULL;
-
-    if (!(key = LDNewObject())) {
-        LD_LOG(LD_LOG_ERROR, "alloc error");
-
-        return NULL;
-    }
-
-    tmp = LDObjectLookup(event, "variation");
-
-    if (LDi_notNull(tmp)) {
-        LD_ASSERT(LDJSONGetType(tmp) == LDNumber);
-
-        if (!(tmp = LDJSONDuplicate(tmp))) {
-            LD_LOG(LD_LOG_ERROR, "alloc error");
-
-            LDJSONFree(key);
-
-            return NULL;
-        }
-
-        if (!LDObjectSetKey(key, "variation", tmp)) {
-            LD_LOG(LD_LOG_ERROR, "alloc error");
-
-            LDJSONFree(key);
-            LDJSONFree(tmp);
-
-            return NULL;
-        }
-    }
-
-    tmp = LDObjectLookup(event, "version");
-
-    if (LDi_notNull(tmp)) {
-        LD_ASSERT(LDJSONGetType(tmp) == LDNumber);
-
-        if (!(tmp = LDJSONDuplicate(tmp))) {
-            LD_LOG(LD_LOG_ERROR, "alloc error");
-
-            LDJSONFree(key);
-
-            return NULL;
-        }
-
-        if (!LDObjectSetKey(key, "version", tmp)) {
-            LD_LOG(LD_LOG_ERROR, "alloc error");
-
-            LDJSONFree(key);
-            LDJSONFree(tmp);
-
-            return NULL;
-        }
-    }
-
-    if (!(keytext = LDJSONSerialize(key))) {
-        LD_LOG(LD_LOG_ERROR, "alloc error");
-
-        LDJSONFree(key);
-
-        return NULL;
-    }
-
-    LDJSONFree(key);
-
-    return keytext;
-}
-
-bool
-LDi_summarizeEvent(struct LDClient *const client,
-    const struct LDJSON *const event, const bool unknown)
-{
-    char *keytext;
-    const char *flagKey;
-    struct LDJSON *tmp, *entry, *flagContext, *counters;
-    bool success;
-
-    LD_ASSERT(client);
-    LD_ASSERT(event);
-
-    keytext     = NULL;
-    flagKey     = NULL;
-    tmp         = NULL;
-    entry       = NULL;
-    flagContext = NULL;
-    counters    = NULL;
-    success     = false;
-
-    tmp = LDObjectLookup(event, "key");
-    LD_ASSERT(tmp);
-    LD_ASSERT(LDJSONGetType(tmp) == LDText);
-    flagKey = LDGetText(tmp);
-    LD_ASSERT(flagKey);
-
-    if (!(keytext = LDi_makeSummaryKey(event))) {
-        LD_LOG(LD_LOG_ERROR, "alloc error");
-
-        return false;
-    }
-
-    LDi_rwlock_wrlock(&client->lock);
-
-    if (client->summaryStart == 0) {
-        unsigned long now;
-
-        LDi_getUnixMilliseconds(&now);
-
-        client->summaryStart = now;
-    }
-
-    if (!(flagContext = LDObjectLookup(client->summaryCounters, flagKey))) {
-        if (!(flagContext = LDNewObject())) {
-            LD_LOG(LD_LOG_ERROR, "alloc error");
-
-            goto cleanup;
-        }
-
-        tmp = LDObjectLookup(event, "default");
-
-        if (LDi_notNull(tmp)) {
-            if (!(tmp = LDJSONDuplicate(tmp))) {
-                LD_LOG(LD_LOG_ERROR, "alloc error");
-
-                LDJSONFree(flagContext);
-
-                goto cleanup;
-            }
-
-            if (!LDObjectSetKey(flagContext, "default", tmp)) {
-                LD_LOG(LD_LOG_ERROR, "alloc error");
-
-                LDJSONFree(tmp);
-                LDJSONFree(flagContext);
-
-                goto cleanup;
-            }
-        }
-
-        if (!(tmp = LDNewObject())) {
-            LD_LOG(LD_LOG_ERROR, "alloc error");
-
-            LDJSONFree(flagContext);
-
-            goto cleanup;
-        }
-
-        if (!LDObjectSetKey(flagContext, "counters", tmp)) {
-            LD_LOG(LD_LOG_ERROR, "alloc error");
-
-            LDJSONFree(tmp);
-            LDJSONFree(flagContext);
-
-            goto cleanup;
-        }
-
-        if (!LDObjectSetKey(client->summaryCounters, flagKey, flagContext)) {
-            LD_LOG(LD_LOG_ERROR, "alloc error");
-
-            LDJSONFree(flagContext);
-
-            goto cleanup;
-        }
-    }
-
-    counters = LDObjectLookup(flagContext, "counters");
-    LD_ASSERT(counters);
-    LD_ASSERT(LDJSONGetType(counters) == LDObject);
-
-    if (!(entry = LDObjectLookup(counters, keytext))) {
-        if (!(entry = LDNewObject())) {
-            LD_LOG(LD_LOG_ERROR, "alloc error");
-
-            goto cleanup;
-        }
-
-        if (!(tmp = LDNewNumber(1))) {
-            LD_LOG(LD_LOG_ERROR, "alloc error");
-
-            LDJSONFree(entry);
-
-            goto cleanup;
-        }
-
-        if (!LDObjectSetKey(entry, "count", tmp)) {
-            LD_LOG(LD_LOG_ERROR, "alloc error");
-
-            LDJSONFree(tmp);
-            LDJSONFree(entry);
-
-            goto cleanup;
-        }
-
-        tmp = LDObjectLookup(event, "value");
-
-        if (LDi_notNull(tmp)) {
-            if (!(tmp = LDJSONDuplicate(tmp))) {
-                LD_LOG(LD_LOG_ERROR, "alloc error");
-
-                LDJSONFree(entry);
-
-                goto cleanup;
-            }
-
-            if (!LDObjectSetKey(entry, "value", tmp)) {
-                LD_LOG(LD_LOG_ERROR, "alloc error");
-
-                LDJSONFree(tmp);
-                LDJSONFree(entry);
-
-                goto cleanup;
-            }
-        }
-
-        tmp = LDObjectLookup(event, "version");
-
-        if (LDi_notNull(tmp)) {
-            if (!(tmp = LDJSONDuplicate(tmp))) {
-                LD_LOG(LD_LOG_ERROR, "alloc error");
-
-                LDJSONFree(entry);
-
-                goto cleanup;
-            }
-
-            if (!LDObjectSetKey(entry, "version", tmp)) {
-                LD_LOG(LD_LOG_ERROR, "alloc error");
-
-                LDJSONFree(tmp);
-                LDJSONFree(entry);
-
-                goto cleanup;
-            }
-        }
-
-        tmp = LDObjectLookup(event, "variation");
-
-        if (LDi_notNull(tmp)) {
-            if (!(tmp = LDJSONDuplicate(tmp))) {
-                LD_LOG(LD_LOG_ERROR, "alloc error");
-
-                LDJSONFree(entry);
-
-                goto cleanup;
-            }
-
-            if (!LDObjectSetKey(entry, "variation", tmp)) {
-                LD_LOG(LD_LOG_ERROR, "alloc error");
-
-                LDJSONFree(tmp);
-                LDJSONFree(entry);
-
-                goto cleanup;
-            }
-        }
-
-        if (unknown) {
-            if (!(tmp = LDNewBool(true))) {
-                LD_LOG(LD_LOG_ERROR, "alloc error");
-
-                LDJSONFree(entry);
-
-                goto cleanup;
-            }
-
-            if (!LDObjectSetKey(entry, "unknown", tmp)) {
-                LD_LOG(LD_LOG_ERROR, "alloc error");
-
-                LDJSONFree(tmp);
-                LDJSONFree(entry);
-
-                goto cleanup;
-            }
-        }
-
-        if (!LDObjectSetKey(counters, keytext, entry)) {
-            LD_LOG(LD_LOG_ERROR, "alloc error");
-
-            LDJSONFree(entry);
-
-            goto cleanup;
-        }
-    } else {
-        bool status;
-        tmp = LDObjectLookup(entry, "count");
-        LD_ASSERT(tmp);
-        status = LDSetNumber(tmp, LDGetNumber(tmp) + 1);
-        LD_ASSERT(status);
-    }
-
-    success = true;
-
-  cleanup:
-    LDi_rwlock_wrunlock(&client->lock);
-
-    LDFree(keytext);
-
-    return success;
 }
 
 struct AnalyticsContext {
@@ -952,150 +113,6 @@ destroy(void *const rawcontext)
     LDFree(context);
 }
 
-static struct LDJSON *
-objectToArray(const struct LDJSON *const object)
-{
-    struct LDJSON *iter, *array;
-
-    LD_ASSERT(object);
-    LD_ASSERT(LDJSONGetType(object) == LDObject);
-
-    iter  = NULL;
-    array = NULL;
-
-    if (!(array = LDNewArray())) {
-        LD_LOG(LD_LOG_ERROR, "alloc error");
-
-        return NULL;
-    }
-
-    for (iter = LDGetIter(object); iter; iter = LDIterNext(iter)) {
-        struct LDJSON *dupe;
-
-        if (!(dupe = LDJSONDuplicate(iter))) {
-            LD_LOG(LD_LOG_ERROR, "alloc error");
-
-            LDJSONFree(array);
-
-            return NULL;
-        }
-
-        LDArrayPush(array, dupe);
-    }
-
-    return array;
-}
-
-struct LDJSON *
-LDi_prepareSummaryEvent(struct LDClient *const client)
-{
-    unsigned long now;
-    struct LDJSON *tmp, *summary, *iter, *counters;
-
-    LD_ASSERT(client);
-
-    tmp      = NULL;
-    summary  = NULL;
-    iter     = NULL;
-    counters = NULL;
-    now      = 0;
-
-    if (!(summary = LDNewObject())) {
-        LD_LOG(LD_LOG_ERROR, "alloc error");
-
-        goto error;
-    }
-
-    if (!(tmp = LDNewText("summary"))) {
-        LD_LOG(LD_LOG_ERROR, "alloc error");
-
-        goto error;
-    }
-
-    if (!(LDObjectSetKey(summary, "kind", tmp))) {
-        LD_LOG(LD_LOG_ERROR, "alloc error");
-
-        LDJSONFree(tmp);
-
-        goto error;
-    }
-
-    if (!(tmp = LDNewNumber(client->summaryStart))) {
-        LD_LOG(LD_LOG_ERROR, "alloc error");
-
-        goto error;
-    }
-
-    if (!(LDObjectSetKey(summary, "startDate", tmp))) {
-        LD_LOG(LD_LOG_ERROR, "alloc error");
-
-        LDJSONFree(tmp);
-
-        goto error;
-    }
-
-    LDi_getUnixMilliseconds(&now);
-
-    if (!(tmp = LDNewNumber(now))) {
-        LD_LOG(LD_LOG_ERROR, "alloc error");
-
-        goto error;
-    }
-
-    if (!(LDObjectSetKey(summary, "endDate", tmp))) {
-        LD_LOG(LD_LOG_ERROR, "alloc error");
-
-        LDJSONFree(tmp);
-
-        goto error;
-    }
-
-    if (!(counters = LDJSONDuplicate(client->summaryCounters))) {
-        LD_LOG(LD_LOG_ERROR, "alloc error");
-
-        goto error;
-    }
-
-    for (iter = LDGetIter(counters); iter; iter = LDIterNext(iter)) {
-        struct LDJSON *countersObject, *countersArray;
-
-        countersObject = LDObjectDetachKey(iter, "counters");
-        LD_ASSERT(countersObject);
-
-        if (!(countersArray = objectToArray(countersObject))) {
-            LD_LOG(LD_LOG_ERROR, "alloc error");
-
-            LDJSONFree(countersObject);
-
-            goto error;
-        }
-
-        LDJSONFree(countersObject);
-
-        if (!LDObjectSetKey(iter, "counters", countersArray)) {
-            LD_LOG(LD_LOG_ERROR, "alloc error");
-
-            LDJSONFree(countersArray);
-
-            goto error;
-        }
-    }
-
-    if (!LDObjectSetKey(summary, "features", counters)) {
-        LD_LOG(LD_LOG_ERROR, "alloc error");
-
-        goto error;
-    }
-
-    return summary;
-
-  error:
-    LDJSONFree(summary);
-    LDJSONFree(counters);
-
-    return NULL;
-}
-
 static const char *
 strnchr(const char *str, const char c, size_t len)
 {
@@ -1172,9 +189,10 @@ LDi_onHeader(const char *buffer, const size_t size,
         return total;
     }
 
-    LDi_rwlock_wrlock(&client->lock);
-    client->lastServerTime = 1000 * (unsigned long long)mktime(&tm);
-    LDi_rwlock_wrunlock(&client->lock);
+    LDi_setServerTime(
+        client->eventProcessor,
+        1000 * (unsigned long long)mktime(&tm)
+    );
 
     return total;
 
@@ -1188,7 +206,6 @@ poll(struct LDClient *const client, void *const rawcontext)
     char url[4096];
     const char *mime, *schema;
     bool shouldFlush;
-    struct LDJSON *summaryEvent;
     bool lastFailed;
 
     LD_ASSERT(rawcontext);
@@ -1218,23 +235,13 @@ poll(struct LDClient *const client, void *const rawcontext)
     }
 
     if (!lastFailed) {
-        struct LDJSON *nextEvents, *nextSummaryCounters;
+        struct LDJSON *events;
 
-        nextEvents          = NULL;
-        nextSummaryCounters = NULL;
+        events = NULL;
 
-        LDi_rwlock_wrlock(&client->lock);
-        if (LDCollectionGetSize(client->events) == 0 &&
-            LDCollectionGetSize(client->summaryCounters) == 0)
-        {
-            LDi_rwlock_wrunlock(&client->lock);
-
-            client->shouldFlush = false;
-
-            return NULL;
-        }
+        LDi_rwlock_rdlock(&client->lock);
         shouldFlush = client->shouldFlush;
-        LDi_rwlock_wrunlock(&client->lock);
+        LDi_rwlock_rdunlock(&client->lock);
 
         if (!shouldFlush) {
             unsigned long now;
@@ -1249,56 +256,26 @@ poll(struct LDClient *const client, void *const rawcontext)
             }
         }
 
-        /* serialize events */
-
-        if (!(nextEvents = LDNewArray())) {
-            LD_LOG(LD_LOG_ERROR, "alloc error");
+        if (!LDi_bundleEventPayload(client->eventProcessor, &events)) {
+            LD_LOG(LD_LOG_ERROR, "failed bundling events");
 
             return NULL;
         }
 
-        if (!(nextSummaryCounters = LDNewObject())) {
-            LD_LOG(LD_LOG_ERROR, "alloc error");
-
-            LDJSONFree(nextEvents);
-
-            return NULL;
-        }
-
-        LDi_rwlock_wrlock(&client->lock);
-
-        if (!(summaryEvent = LDi_prepareSummaryEvent(client))) {
-            LD_LOG(LD_LOG_ERROR, "failed to prepare summary");
-
+        if (!events) {
+            /* no events to send */
+            LDi_rwlock_wrlock(&client->lock);
+            shouldFlush = false;
             LDi_rwlock_wrunlock(&client->lock);
 
-            LDJSONFree(nextEvents);
-            LDJSONFree(nextSummaryCounters);
-
             return NULL;
         }
 
-        LDArrayPush(client->events, summaryEvent);
-
-        if (!(context->buffer = LDJSONSerialize(client->events))) {
+        if (!(context->buffer = LDJSONSerialize(events))) {
             LD_LOG(LD_LOG_ERROR, "alloc error");
 
-            LDi_rwlock_wrunlock(&client->lock);
-
-            LDJSONFree(nextEvents);
-            LDJSONFree(nextSummaryCounters);
-
             return NULL;
         }
-
-        LDJSONFree(client->events);
-        LDJSONFree(client->summaryCounters);
-
-        client->summaryStart    = 0;
-        client->events          = nextEvents;
-        client->summaryCounters = nextSummaryCounters;
-
-        LDi_rwlock_wrunlock(&client->lock);
 
         /* Only generate a UUID once per payload. We want the header to remain
         the same during a retry */
